Move form to ContactForm component
